# openapi/lis-internal-v1.yaml
openapi: 3.1.0
info:
  title: LIS Core API
  summary: Starter modular monolith for a laboratory information system
  description: Internal API with working persistence for auth, master data, orders,
    specimen lifecycle, task orchestration, observations, reports, audit, provenance,
    device registry, HL7 v2 integrations, device gateway, autoverification, ASTM-style
    drivers, and a read/search FHIR facade.
  version: 0.7.0
paths:
...
  /fhir/R4/Patient:
    get:
      tags:
      - fhir
      summary: Search Patients
      operationId: search_patients_fhir_R4_Patient_get
      security:
      - HTTPBearer: []
      parameters:
      - name: identifier
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Identifier
      - name: family
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Family
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Patients Fhir R4 Patient Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/Patient/{patient_id}:
    get:
      tags:
      - fhir
      summary: Read Patient
      operationId: read_patient_fhir_R4_Patient__patient_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Patient Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Read Patient Fhir R4 Patient  Patient Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/patients:
    post:
      tags:
      - patients
      summary: Create Patient
      operationId: create_patient_api_v1_patients_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - patients
      summary: List Patients
      operationId: list_patients_api_v1_patients_get
      security:
      - HTTPBearer: []
      parameters:
      - name: mrn
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Mrn
      - name: family_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Family Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/PatientSummary'
                title: Response List Patients Api V1 Patients Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{patient_id}:
    get:
      tags:
      - patients
      summary: Get Patient
      operationId: get_patient_api_v1_patients__patient_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Patient Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/reports/generate:
    post:
      tags:
      - reports
      summary: Generate Report
      operationId: generate_report_api_v1_reports_generate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosticReportSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/reports/{report_id}:
    get:
      tags:
      - reports
      summary: Get Report
      operationId: get_report_api_v1_reports__report_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Report Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosticReportSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reports/{report_id}/authorize:
    post:
      tags:
      - reports
      summary: Authorize Report
      operationId: authorize_report_api_v1_reports__report_id__authorize_post
      security:
      - HTTPBearer: []
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Report Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizeReportRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosticReportSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reports/{report_id}/amend:
    post:
      tags:
      - reports
      summary: Amend Report
      operationId: amend_report_api_v1_reports__report_id__amend_post
      security:
      - HTTPBearer: []
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Report Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmendReportRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosticReportSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reports/{report_id}/pdf:
    get:
      tags:
      - reports
      summary: Get Report Pdf
      operationId: get_report_pdf_api_v1_reports__report_id__pdf_get
      security:
      - HTTPBearer: []
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Report Id
      - name: version
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 1
          - type: 'null'
          title: Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
            application/pdf: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AmendReportRequest:
      properties:
        signed_by_user_id:
          type: string
          format: uuid
          title: Signed By User Id
        reason:
          type: string
          title: Reason
        conclusion_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Conclusion Text
      additionalProperties: false
      type: object
      required:
      - signed_by_user_id
      - reason
      title: AmendReportRequest
...
    AuthorizeReportRequest:
      properties:
        signed_by_user_id:
          type: string
          format: uuid
          title: Signed By User Id
      additionalProperties: false
      type: object
      required:
      - signed_by_user_id
      title: AuthorizeReportRequest
...
    CreateReportRequest:
      properties:
        order_id:
          type: string
          format: uuid
          title: Order Id
        code_local:
          anyOf:
          - type: string
          - type: 'null'
          title: Code Local
        code_loinc:
          anyOf:
          - type: string
          - type: 'null'
          title: Code Loinc
        conclusion_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Conclusion Text
      additionalProperties: false
      type: object
      required:
      - order_id
      title: CreateReportRequest
...
    DiagnosticReportSummary:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        report_no:
          type: string
          title: Report No
        order_id:
          type: string
          format: uuid
          title: Order Id
        patient_id:
          type: string
          format: uuid
          title: Patient Id
        status:
          $ref: '#/components/schemas/ReportStatus'
        category_code:
          type: string
          title: Category Code
        code_local:
          anyOf:
          - type: string
          - type: 'null'
          title: Code Local
        code_loinc:
          anyOf:
          - type: string
          - type: 'null'
          title: Code Loinc
        effective_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Effective At
        issued_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Issued At
        conclusion_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Conclusion Text
        current_version_no:
          type: integer
          title: Current Version No
        versions:
          items:
            $ref: '#/components/schemas/DiagnosticReportVersionSummary'
          type: array
          title: Versions
        observation_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Observation Ids
      additionalProperties: false
      type: object
      required:
      - id
      - report_no
      - order_id
      - patient_id
      - status
      - category_code
      - current_version_no
      title: DiagnosticReportSummary
    DiagnosticReportVersionSummary:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        report_id:
          type: string
          format: uuid
          title: Report Id
        version_no:
          type: integer
          title: Version No
        status:
          $ref: '#/components/schemas/ReportStatus'
        amendment_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Amendment Reason
        rendered_pdf_uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Rendered Pdf Uri
        signed_by_user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Signed By User Id
        signed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Signed At
        payload:
          additionalProperties: true
          type: object
          title: Payload
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - report_id
      - version_no
      - status
      - created_at
      title: DiagnosticReportVersionSummary
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PatientCreateRequest:
      properties:
        mrn:
          type: string
          title: Mrn
        given_name:
          type: string
          title: Given Name
        family_name:
          type: string
          title: Family Name
        sex_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Sex Code
        birth_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Birth Date
      additionalProperties: false
      type: object
      required:
      - mrn
      - given_name
      - family_name
      title: PatientCreateRequest
    PatientSummary:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        mrn:
          type: string
          title: Mrn
        given_name:
          type: string
          title: Given Name
        family_name:
          type: string
          title: Family Name
        sex_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Sex Code
        birth_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Birth Date
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - mrn
      - given_name
      - family_name
      - created_at
      title: PatientSummary
...
    ReportStatus:
      type: string
      enum:
      - registered
      - partial
      - preliminary
      - final
      - amended
      - corrected
      - entered_in_error
      title: ReportStatus
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer